requirements.txt
aioredis==1.0.0
asgiref==2.1.6
asn1crypto==0.23.0
aspectlib==1.4.2
async-timeout==2.0.0
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
certifi==2017.11.5
cffi==1.11.2
channels==2.0.2
channels-redis==2.1.0
chardet==3.0.4
constantly==15.1.0
daphne==2.0.4
Django==2.0.2
django-bootstrap3==9.1.0
django-cors-middleware==1.3.1
django-extensions==1.9.1
django-model-utils==3.1.1
django-oauth-toolkit==1.0.0
fields==5.0.0
hiredis==0.2.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
msgpack==0.5.6
msgpack-python==0.5.1
oauthlib==2.0.6
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.1
pycparser==2.18
pytz==2018.3
redis==2.10.6
requests==2.18.4
shortuuid==0.5.0
six==1.11.0
style==1.1.0
Twisted==17.9.0
txaio==2.8.2
Unipath==1.1
urllib3==1.22
zope.interface==4.4.3
